backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.6.0
python-dotenv==1.0.1
pydantic==2.8.2
python-multipart==0.0.6
